+++ packages/card/src/components/CheckCard/index.tsx
@@ -1,6 +1,5 @@
 import { useMountMergeState } from '@ant-design/pro-utils';
 import { Avatar, ConfigProvider } from 'antd';
-
 import classNames from 'classnames';
 import type { MouseEventHandler } from 'react';
 import React, { useContext, useEffect, useMemo } from 'react';
@@ -8,128 +7,152 @@ import ProCardActions from '../Actions';
 import type { CheckCardGroupProps } from './Group';
 import CheckCardGroup, { CardLoading, CheckCardGroupConnext } from './Group';
 import { useStyle } from './style';
+
+/**
+ * Props for the CheckCard component.
+ */
 interface CheckCardProps {
   /**
-   * 自定义前缀
+   * Custom prefix class.
    *
    * @ignore
    */
   prefixCls?: string;
-  /** Change 回调 */
+  /**
+   * Callback function for change event.
+   *
+   * @param checked - The checked state.
+   */
   onChange?: (checked: boolean) => void;
-  /** Click 回调 */
+  /**
+   * Callback function for click event.
+   *
+   * @param event - The click event.
+   */
   onClick?: (event: MouseEventHandler<HTMLDivElement> | undefined) => void;
-  /** 鼠标进入时的回调 */
+  /**
+   * Callback function for mouse enter event.
+   *
+   * @param event - The mouse enter event.
+   */
   onMouseEnter?: MouseEventHandler<HTMLDivElement>;
-  /** 鼠标出来时的回调 */
+  /**
+   * Callback function for mouse leave event.
+   *
+   * @param event - The mouse leave event.
+   */
   onMouseLeave?: (event: MouseEventHandler<HTMLDivElement> | undefined) => void;
   /**
-   * 默认是否勾选
+   * Whether the card is initially checked.
    *
    * @default false
-   * @title 默认勾选
+   * @title Default Checked
    */
   defaultChecked?: boolean;
   /**
-   * 强制勾选
+   * Whether the card is checked.
    *
    * @default false
-   * @title 强制勾选
+   * @title Checked
    */
   checked?: boolean;
   /**
-   * 不可用
+   * Whether the card is disabled.
    *
    * @default false
-   * @title 禁用
+   * @title Disabled
    */
   disabled?: boolean;
   /**
-   * 选项卡样式
+   * Custom style for the card.
    *
    * @ignore
    */
   style?: React.CSSProperties;
   /**
-   * 选项卡 className
+   * Custom class name for the card.
    *
    * @ignore
    */
   className?: string;
   /**
-   * 左侧头像展示，可以是一个链接也可以是是一个 ReactNode
+   * The avatar to display on the left side, can be a link or a ReactNode.
    *
-   * @title 头像
+   * @title Avatar
    */
   avatar?: React.ReactNode;
   /**
-   * 标题展示
+   * The title to display.
    *
-   * @title 标题
+   * @title Title
    */
   title?: React.ReactNode;
   /**
-   * 二级标题展示
+   * The subtitle to display.
    *
-   * @title 二级标题
+   * @title Subtitle
    */
   subTitle?: React.ReactNode;
   /**
-   * 描述展示
+   * The description to display.
    *
-   * @title 描述
+   * @title Description
    */
   description?: React.ReactNode;
   /**
-   * 选项值
+   * The value of the card.
    *
-   * @title 值
+   * @title Value
    */
   value?: any;
   /**
-   * 内容是否在加载中
+   * Whether the content is in loading state.
    *
    * @default false
-   * @title 加载中
+   * @title Loading
    */
   loading?: boolean;
   /**
-   * 图片封面默认，该模式下其他展示值被忽略
+   * The cover image of the card. Other display values are ignored in this mode.
    *
-   * @title 卡片背景图片
+   * @title Card Background Image
    */
   cover?: React.ReactNode;
   /**
-   * 组件尺寸，支持大，中，小三种默认尺寸，用户可以自定义宽高
+   * The size of the component. Supports 'large', 'default', and 'small' sizes. Users can also customize the width and height.
    *
    * @default default
-   * @title 选择框大小
+   * @title Checkbox Size
    */
   size?: 'large' | 'default' | 'small';
   /**
-   * 是否显示边框
+   * Whether to show the border.
    *
    * @default true
-   * @title 显示边框
+   * @title Show Border
    */
   bordered?: boolean;
   /**
-   * 卡片右上角的操作区域
+   * The action area in the top right corner of the card.
    *
-   * @title 操作栏
+   * @title Actions
    */
   extra?: React.ReactNode;
-
+  /**
+   * The content of the card.
+   */
   children?: React.ReactNode;
   /**
-   * 内容区域的样式设计
+   * Custom style for the content area.
    */
   bodyStyle?: React.CSSProperties;
   /**
-   * 右下角的操作区
+   * The action area in the bottom right corner.
    */
   actions?: React.ReactNode[];
-
+  /**
+   * Whether the card is in ghost mode.
+   */
   ghost?: boolean;
 }
 
@@ -170,25 +193,6 @@ const CheckCard: React.FC<CheckCardProps> & {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [props.value]);
 
-  /**
-   * 头像自定义
-   *
-   * @param prefixCls
-   * @param cover
-   * @returns
-   */
-  const renderCover = (prefixCls: string, cover: string | React.ReactNode) => {
-    return (
-      <div className={`${prefixCls}-cover`}>
-        {typeof cover === 'string' ? (
-          <img src={cover} alt="checkcard" />
-        ) : (
-          cover
-        )}
-      </div>
-    );
-  };
-
   const {
     prefixCls: customizePrefixCls,
     className,
@@ -205,6 +209,25 @@ const CheckCard: React.FC<CheckCardProps> & {
   const prefixCls = getPrefixCls('pro-checkcard', customizePrefixCls);
 
   const { wrapSSR, hashId } = useStyle(prefixCls);
+
+  /**
+   * 头像自定义
+   *
+   * @param cls
+   * @param coverDom
+   * @returns
+   */
+  const renderCover = (cls: string, coverDom: string | React.ReactNode) => {
+    return (
+      <div className={classNames(`${cls}-cover`, hashId)}>
+        {typeof coverDom === 'string' ? (
+          <img src={coverDom} alt="checkcard" />
+        ) : (
+          coverDom
+        )}
+      </div>
+    );
+  };
 
   checkCardProps.checked = stateChecked;
 
@@ -248,7 +271,7 @@ const CheckCard: React.FC<CheckCardProps> & {
 
   const metaDom = useMemo(() => {
     if (cardLoading) {
-      return <CardLoading prefixCls={prefixCls || ''} />;
+      return <CardLoading prefixCls={prefixCls || ''} hashId={hashId} />;
     }
 
     if (cover) {
@@ -256,7 +279,7 @@ const CheckCard: React.FC<CheckCardProps> & {
     }
 
     const avatarDom = avatar ? (
-      <div className={`${prefixCls}-avatar ${hashId}`.trim()}>
+      <div className={classNames(`${prefixCls}-avatar`, hashId)}>
         {typeof avatar === 'string' ? (
           <Avatar size={48} shape="square" src={avatar} />
         ) : (
@@ -266,23 +289,27 @@ const CheckCard: React.FC<CheckCardProps> & {
     ) : null;
 
     const headerDom = (title ?? extra) != null && (
-      <div className={`${prefixCls}-header ${hashId}`.trim()}>
-        <div className={`${prefixCls}-header-left ${hashId}`.trim()}>
-          <div className={`${prefixCls}-title ${hashId}`.trim()}>{title}</div>
+      <div className={classNames(`${prefixCls}-header`, hashId)}>
+        <div className={classNames(`${prefixCls}-header-left`, hashId)}>
+          <div className={classNames(`${prefixCls}-title`, hashId)}>
+            {title}
+          </div>
           {props.subTitle ? (
-            <div className={`${prefixCls}-subTitle ${hashId}`.trim()}>
+            <div className={classNames(`${prefixCls}-subTitle`, hashId)}>
               {props.subTitle}
             </div>
           ) : null}
         </div>
         {extra && (
-          <div className={`${prefixCls}-extra ${hashId}`.trim()}>{extra}</div>
+          <div className={classNames(`${prefixCls}-extra`, hashId)}>
+            {extra}
+          </div>
         )}
       </div>
     );
 
     const descriptionDom = description ? (
-      <div className={`${prefixCls}-description ${hashId}`.trim()}>
+      <div className={classNames(`${prefixCls}-description`, hashId)}>
         {description}
       </div>
     ) : null;
@@ -295,7 +322,7 @@ const CheckCard: React.FC<CheckCardProps> & {
       <div className={metaClass}>
         {avatarDom}
         {headerDom || descriptionDom ? (
-          <div className={`${prefixCls}-detail ${hashId}`.trim()}>
+          <div className={classNames(`${prefixCls}-detail`, hashId)}>
             {headerDom}
             {descriptionDom}
           </div>
@@ -328,7 +355,7 @@ const CheckCard: React.FC<CheckCardProps> & {
       {metaDom}
       {props.children ? (
         <div
-          className={classNames(`${prefixCls}-body`)}
+          className={classNames(`${prefixCls}-body`, hashId)}
           style={props.bodyStyle}
         >
           {props.children}