+++ packages/form/src/components/LoginForm/demos/login-form.tsx
@@ -36,7 +36,7 @@ export default () => {
     <ProConfigProvider hashed={false}>
       <div style={{ backgroundColor: token.colorBgContainer }}>
         <LoginForm
-          logo="https://github.githubassets.com/images/modules/logos_page/Octocat.png"
+          logo="https://github.githubassets.com/favicons/favicon.png"
           title="Github"
           subTitle="全球最大的代码托管平台"
           actions={