+++ packages/form/src/layouts/ModalForm/index.tsx
@@ -6,7 +6,6 @@ import { noteOnce } from 'rc-util/lib/warning';
 import React, {
   useCallback,
   useContext,
-  useEffect,
   useImperativeHandle,
   useMemo,
   useRef,
@@ -54,7 +53,7 @@ export type ModalFormProps<
     visible?: boolean;
 
     /** @name 打开关闭的事件 */
-    onOpenChange?: (visible: boolean) => void;
+    onOpenChange?: (open: boolean) => void;
     /**
      * 不支持 'visible'，请使用全局的 visible
      *
@@ -129,13 +128,14 @@ function ModalForm<T = Record<string, any>, U = Record<string, any>>({
     [formRef.current],
   );
 
-  useEffect(() => {
-    if (open && (propsOpen || propVisible)) {
-      onOpenChange?.(true);
-      onVisibleChange?.(true);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [propVisible, propsOpen, open]);
+  // useMergedState监听了open的变化
+  // useEffect(() => {
+  //   if (open && (propsOpen || propVisible)) {
+  //     onOpenChange?.(true);
+  //     onVisibleChange?.(true);
+  //   }
+  //   // eslint-disable-next-line react-hooks/exhaustive-deps
+  // }, [propVisible, propsOpen, open]);
 
   const triggerDom = useMemo(() => {
     if (!trigger) {
@@ -243,7 +243,9 @@ function ModalForm<T = Record<string, any>, U = Record<string, any>>({
         }}
         afterClose={() => {
           resetFields();
-          setOpen(false);
+          if (open) {
+            setOpen(false);
+          }
           modalProps?.afterClose?.();
         }}
         footer={