+++ packages/provider/src/typing/layoutToken.ts
@@ -15,7 +15,6 @@ export type BaseLayoutDesignToken = {
    * 跨站点应用的图标hover颜色
    */
   colorTextAppListIcon: string;
-
   /**
    * layout 的背景颜色
    */