+++ tests/form/modalForm.test.tsx
@@ -266,26 +266,27 @@ describe('ModalForm', () => {
     await waitFor(async () => {
       await waitForWaitTime(100);
     });
-    expect(fn).toHaveBeenCalledWith(false);
-    expect(fn).toHaveBeenCalledTimes(2);
+    expect(fn).toBeCalledWith(false);
+    expect(fn).toBeCalledTimes(1); // 关闭只触发一次 onOpenChange
   });
 
-  it('📦 modal open=true simulate onOpenChange', async () => {
-    const fn = vi.fn();
-    render(
-      <ModalForm
-        open
-        trigger={<Button id="new">新建</Button>}
-        onOpenChange={(visible) => fn(visible)}
-      >
-        <ProFormText name="name" />
-      </ModalForm>,
-    );
+  // 如果默认就是 open 的 onOpenChange 预期应当不触发
+  // it('📦 modal open=true simulate onOpenChange', async () => {
+  //   const fn = vi.fn();
+  //   render(
+  //     <ModalForm
+  //       open
+  //       trigger={<Button id="new">新建</Button>}
+  //       onOpenChange={(visible) => fn(visible)}
+  //     >
+  //       <ProFormText name="name" />
+  //     </ModalForm>,
+  //   );
 
-    await waitFor(() => {
-      expect(fn).toHaveBeenCalledWith(true);
-    });
-  });
+  //   await waitFor(() => {
+  //     expect(fn).toBeCalledWith(true);
+  //   });
+  // });
 
   it('📦 reset button will simulate onOpenChange', async () => {
     const fn = vi.fn();
@@ -385,7 +386,7 @@ describe('ModalForm', () => {
     await waitFor(async () => {
       await waitForWaitTime(100);
     });
-    expect(fn).toHaveBeenCalledTimes(1);
+    expect(fn).toBeCalledTimes(0);
   });
 
   it('📦 ModalForm support submitter is false', async () => {