+++ packages/components/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.7.5...@ant-design/pro-components@2.7.6) (2024-04-20)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.7.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.7.3...@ant-design/pro-components@2.7.5) (2024-04-20)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.7.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.7.3...@ant-design/pro-components@2.7.4) (2024-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.7.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.7.2...@ant-design/pro-components@2.7.3) (2024-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.7.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.7.1...@ant-design/pro-components@2.7.2) (2024-04-17)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.7.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.7.0...@ant-design/pro-components@2.7.1) (2024-04-17)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
 # [2.7.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.6.52...@ant-design/pro-components@2.7.0) (2024-03-27)
 
 ### Features