+++ packages/list/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.57](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.56...@ant-design/pro-list@2.5.57) (2024-04-20)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.5.56](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.54...@ant-design/pro-list@2.5.56) (2024-04-20)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.5.55](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.54...@ant-design/pro-list@2.5.55) (2024-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.5.54](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.53...@ant-design/pro-list@2.5.54) (2024-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.5.53](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.52...@ant-design/pro-list@2.5.53) (2024-04-17)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [2.5.52](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.51...@ant-design/pro-list@2.5.52) (2024-04-17)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [2.5.51](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.50...@ant-design/pro-list@2.5.51) (2024-03-27)
 
 **Note:** Version bump only for package @ant-design/pro-list