+++ packages/provider/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.14.2...@ant-design/pro-provider@2.14.4) (2024-04-20)
+
+**Note:** Version bump only for package @ant-design/pro-provider
+
+## [2.14.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.14.2...@ant-design/pro-provider@2.14.3) (2024-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-provider
+
+## [2.14.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.14.1...@ant-design/pro-provider@2.14.2) (2024-04-17)
+
+**Note:** Version bump only for package @ant-design/pro-provider
+
 ## [2.14.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@2.14.0...@ant-design/pro-provider@2.14.1) (2024-03-27)
 
 ### Bug Fixes