+++ packages/skeleton/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.13](https://github.com/ant-design/pro-components/compare/@ant-design/pro-skeleton@2.1.12...@ant-design/pro-skeleton@2.1.13) (2024-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-skeleton
+
+## [2.1.12](https://github.com/ant-design/pro-components/compare/@ant-design/pro-skeleton@2.1.11...@ant-design/pro-skeleton@2.1.12) (2024-04-17)
+
+**Note:** Version bump only for package @ant-design/pro-skeleton
+
 ## [2.1.11](https://github.com/ant-design/pro-components/compare/@ant-design/pro-skeleton@2.1.10...@ant-design/pro-skeleton@2.1.11) (2024-03-15)
 
 ### Bug Fixes